Problem: a commitment might not be an account Solution: move commitment enum to Machine
